Move platform/jwt into infra/auth
Change-type: patch
