+++ src/infra/auth/auth.ts
@@ -4,7 +4,7 @@ import * as _ from 'lodash';
 import { sbvrUtils, hooks, permissions, errors } from '@balena/pinejs';
 
 import { retrieveAPIKey } from '../../platform/api-keys';
-import { User } from '../../platform/jwt';
+import { User } from './jwt-passport';
 
 import { getIP } from '../../lib/utils';
 import type { User as DbUser } from '../../models';
